storage_manager-0.1.4-Apache-forecast

+++ StorageManager/src/main/scala/com/boke/soft/dsj/bean/MaterialQuantityInfo.scala
@@ -5,5 +5,5 @@ case class MaterialQuantityInfo(item_cd: String, // 物料编码
                                 insert_datetime: String, // 出库日期
                                 quantity: Double, // 出库量
                                 var status: String = null, // 状态
-                                var upper: Int = 0 //每个状态段的数值上限
+                                var upper: Long = 0L //每个状态段的数值上限
                                )
